requirements.txt
python-dotenv
langchain_community
langchain_core


# pip install langchain_community -i https://pypi.tuna.tsinghua.edu.cn/simple